feat : factory 파일 구조 변경
