+++ question-service/src/main/java/com/exit/question/service/util/factory/ResponseCommentFactory.java
@@ -1,50 +1,50 @@
-package com.exit.question.service.util;
+package com.exit.question.service.util.factory;
 
 import com.exit.common.exception.grpc.GrpcException;
 import com.exit.question.controller.dto.request.NotificationContentDto;
-import com.exit.question.controller.dto.request.SendNotificationRequestDto;
 import com.exit.question.domain.Comment;
-import com.exit.question.domain.question.Question;
-import com.exit.question.domain.question.QuestionComment;
-import com.exit.question.domain.question.repository.QuestionCommentRepository;
-import com.exit.question.domain.question.repository.QuestionRepository;
+import com.exit.question.domain.response.Response;
+import com.exit.question.domain.response.ResponseComment;
+import com.exit.question.domain.response.repository.ResponseCommentRepository;
+import com.exit.question.domain.response.repository.ResponseRepository;
 import com.exit.question.exception.GrpcCommentErrorCode;
 import com.exit.question.exception.GrpcQuestionErrorCode;
+import com.exit.question.exception.GrpcResponseErrorCode;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor(access = AccessLevel.PROTECTED)
-public class QuestionCommentFactory extends CommentFactory {
-    private final QuestionRepository questionRepository;
-    private final QuestionCommentRepository questionCommentRepository;
+public class ResponseCommentFactory extends CommentFactory {
+    private final ResponseRepository responseRepository;
+    private final ResponseCommentRepository responseCommentRepository;
 
     @Override
     public Comment createAndSaveComment(Long targetId, Long authorId, String content) {
-        Question question = questionRepository.findById(targetId)
+        Response response = responseRepository.findById(targetId)
                 .orElseThrow(() -> new GrpcException(GrpcQuestionErrorCode.NULL_QUESTION));
 
-        QuestionComment comment = QuestionComment.builder()
-                .question(question)
+        ResponseComment comment = ResponseComment.builder()
+                .response(response)
                 .writerId(authorId)
                 .content(content)
                 .build();
 
-        return questionCommentRepository.save(comment);
+        return responseCommentRepository.save(comment);
     }
 
     @Override
     public void deleteComment(Long targetId, Long authorId) {
         validateCommentWriter(targetId, authorId);
-        questionCommentRepository.deleteById(targetId);
+        responseCommentRepository.deleteById(targetId);
     }
 
     @Override
     public SendNotificationRequestDto createSendNotificationRequestDto(Long targetId, String deviceId) {
-        NotificationContentDto dto = questionRepository.findContentById(targetId)
-                .orElseThrow(() -> new GrpcException(GrpcQuestionErrorCode.NULL_QUESTION));
 
+        NotificationContentDto dto = responseRepository.findContentById(targetId)
+                .orElseThrow(() -> new GrpcException(GrpcResponseErrorCode.NULL_RESPONSE));
         return SendNotificationRequestDto.builder()
                 .body(dto.content())
                 .type("NEW_COMMENT")
@@ -55,7 +55,7 @@ public class QuestionCommentFactory extends CommentFactory {
     }
 
     private void validateCommentWriter(Long commentId, Long userId) {
-        boolean isAuthorized = questionCommentRepository.existsByIdAndWriterId(commentId, userId);
+        boolean isAuthorized = responseCommentRepository.existsByIdAndWriterId(commentId, userId);
         if (!isAuthorized) {
             throw new GrpcException(GrpcCommentErrorCode.COMMENT_WRITER_MISMATCH);
         }
